requirements.txt
Flask==3.0.3
python-dotenv==1.0.0
requests==2.31.0
openai
anthropic
flask-cors==4.0.0
pydantic
typing-extensions
flasgger==0.9.7.1
flask-swagger-ui==5.21.0
pytest==8.3.3
pytest-mock==3.12.0
pytest-flask==1.3.0
pytest-cov==4.1.0
pytest-asyncio==0.24.0
langchain==0.1.20
langchain-community==0.0.38
langchain-openai==0.1.7
langchain-anthropic==0.1.11
faiss-cpu==1.8.0
sentence-transformers==2.7.0
tiktoken==0.7.0